Fix infinite loop in scheduler discovery process
Reading the comments from "icecream/comm.h", it seems we have to call
DiscoverSched::try_get_scheduler() repeatedly while it returns NULL,
as long as it has not timed out. Also, we try to yield in the coroutine
by waiting on input on DiscoverSched::listen_fd() as recommended, as long
as the method returns a valid file descriptor.