autoconf: Compile, not preprocess, for header checks
Essentially, running the preprocessor is terribly broken with certain
combinations of CPPFLAGS and CXXFLAGS, because only CPPFLAGS is passed
down to the preprocessor. See this e-mail thread for a discussion:

  https://lists.gnu.org/archive/html/autoconf/2013-05/msg00006.html

The TL;DR is that using AC_COMPILE_IFELSE always works, because all of
CPPFLAGS and CXXFLAGS get passed to the compiler, and anyway one wants
to ensure that the header can be successfully parsed by the compiler.
Only preprocessing it is as good as only testing for the existence of
the file.